move users.py to models package
