--- backend/ops_api/ops/user/urls.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint("users", __name__, url_prefix="/users")
-
-# bp.add_url_rule("/", view_func=views.get_all_users)
